cleaning up bars
