+++ .github/workflows/codspeed.yml
@@ -33,4 +33,4 @@ jobs:
         uses: CodSpeedHQ/action@v3
         with:
           token: ${{ secrets.CODSPEED_TOKEN }}
-          run: uv run pytest --codspeed
+          run: uv run pytest --codspeed -m 'benchmark and not pandas'
