+++ pyproject.toml
@@ -122,6 +122,7 @@ addopts = [
     "--cov-report=term-missing",
     "--cov-fail-under=80",
     "--cov-config=pyproject.toml",
+    "-m not (pandas or benchmark)",
 ]
 [tool.coverage.run]
 omit = ["polars_trading/_testing/*"]
